feat: repository & dto

+++ src/main/java/alfarezyyd/asclepius/repository/EthnicityRepository.java
@@ -3,5 +3,5 @@ package alfarezyyd.asclepius.repository;
 import alfarezyyd.asclepius.model.entity.Ethnicity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EthnicityRepository extends JpaRepository<Ethnicity, Long> {
+public interface EthnicityRepository extends JpaRepository<Ethnicity, Short> {
 }
