+++ src/main/java/alfarezyyd/asclepius/repository/LanguageRepository.java
@@ -3,5 +3,5 @@ package alfarezyyd.asclepius.repository;
 import alfarezyyd.asclepius.model.entity.Language;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LanguageRepository extends JpaRepository<Language, Long> {
+public interface LanguageRepository extends JpaRepository<Language, Short> {
 }
